+++ packages/thumbprint-react/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Security
+
+-   [Patch] Update version of lodash
+
 ## 14.13.0 - 2022-08-11
 
 ### Added